[TASK] Update kernel modules to support 64 bit addressing constraints
This commit temporarily adds support to handle 64/32 bit system together.
The shared structures are made to use fixed data size variables to remove
platform dependencies and changes in size.

+++ stack/src/kernel/event/eventkcalintf-circbuf.c
@@ -312,9 +312,9 @@ tOplkError eventkcal_processEventCircbuf(tEventQueue eventQueue_p)
     pEvent->eventArgSize = (readSize - sizeof(tEvent));
 
     if (pEvent->eventArgSize > 0)
-        pEvent->pEventArg = &aRxBuffer_l[eventQueue_p][sizeof(tEvent)];
+        pEvent->pEventArg = (ULONGLONG)&aRxBuffer_l[eventQueue_p][sizeof(tEvent)];
     else
-        pEvent->pEventArg = NULL;
+        pEvent->pEventArg = (ULONGLONG)NULL;
 
     /*TRACE("Process Kernel  type:%s(%d) sink:%s(%d) size:%d!\n",
            debugstr_getEventTypeStr(pEvent->eventType), pEvent->eventType,
